requirements.txt
pytest
pytest-asyncio
coverage
redis
pydantic
redis_om
networkx
matplotlib